backend/requirements.txt
Flask==2.2.3
flask-cors==3.1.1
opensearch==2.8.0
sentence-transformers==2.2.2
